ディレクトリ関連のロジックを分離した
